Add files via upload
It is a bit messy, ignore the first few lines of code. Main implementation starts from NASA Api integration please look into that. Performed the wind speed mapping to the turbine factors but facing a minor issue known as (AVERAGE SPEED BELOW CUT IN SPEED GIVES 0 Output energy generated for that Day) Need to fix it.